rename: store 폴더 하나로 병합 및 네이밍 컨벤션 통일
