+++ src/app/pockets/page.tsx
@@ -2,7 +2,7 @@
 
 import React, { useCallback, useEffect, useRef, useState } from "react";
 import { useRouter } from "next/navigation";
-import usePocketStore from "@/app/store/usePocket";
+import usePocketStore from "@/store/pocket";
 import { fetchLuckyPouches } from "@/api/api-form";
 import Notice from "@/components/modal/notice";
 import { validatePouches, Pouch } from "@/utils/validation/validationPouch";
