+++ src/store/result.ts
